Send out a message if encode type not supported.  Change from int to unsigned short.  Check limits.

+++ gsm_essentials.h
@@ -13,7 +13,7 @@ void Parse_ServerList (size_t fileSize, char *fileBuffer, char *gamenameFromHttp
 void AddServers_From_List_Execute (char *fileBuffer, char *gamenameFromHttp); /* FS: From Quake 2 */
 
 /* FS: Aluigi.org stuff */
-unsigned char *gsseckey (unsigned char *dst, unsigned char *src, unsigned char *key, int enctype);
+unsigned char *gsseckey (unsigned char *dst, unsigned char *src, unsigned char *key, unsigned short enctype);
 int create_enctype1_buffer (const char *validate_key, char *input, int inputLen, char *output);
 
 #endif // _GSM_ESSENTIALS_H
